+++ src/Shipment/Models/ShippingMethod.php
@@ -85,6 +85,16 @@ class ShippingMethod extends Model implements ShippingMethodContract
         return $this->getCalculator()->calculate($subject, $this->configuration());
     }
 
+    public function isZoneRestricted(): bool
+    {
+        return null !== $this->zone_id;
+    }
+
+    public function isNotZoneRestricted(): bool
+    {
+        return !$this->isZoneRestricted();
+    }
+
     public function zone(): BelongsTo
     {
         return $this->belongsTo(ZoneProxy::modelClass(), 'zone_id', 'id');